requirements.txt
# MkDocs and Material Theme
mkdocs>=1.5.0
mkdocs-material>=9.5.0

# Plugins for enhanced functionality
mkdocs-git-revision-date-localized-plugin>=1.2.0

# Python dependencies for aggregation script
PyYAML>=6.0
